fix(catalog): isolate publisher validation dependencies
